+++ CHANGELOG.md
@@ -1,8 +1,10 @@
-## 2.12.1 (Not published yet)
+## 2.12.1
 
 ### `@liveblocks/react-ui`
 
 - Prevent unsupported attachment previews from loading infinitely.
+- Refactored `Thread` and `Comment` component to be used outside of the
+  `RoomProvider` component.
 
 ## 2.12.0